Prevent potential overflow on oil pressure protection time. Prevents protection never turning off

+++ speeduino/engineProtection.cpp
@@ -109,7 +109,7 @@ byte checkOilPressureLimit(void)
         if(oilProtStartTime == 0) { oilProtStartTime = div100(millis()); }
 
         /* Check if countdown has reached its target, if so then instruct to cut */
-        if( (uint8_t(div100(millis())) >= (uint16_t(oilProtStartTime + configPage10.oilPressureProtTime)) ) || (alreadyActive > 0) )
+        if( (uint8_t(div100(millis())) >= (uint16_t((uint16_t)oilProtStartTime + configPage10.oilPressureProtTime)) ) || (alreadyActive > 0) )
         {
           BIT_SET(currentStatus.engineProtectStatus, ENGINE_PROTECT_BIT_OIL);
           oilProtectActive = 1;
